# docs/docs/api-reference/openapi.json
{
  "components": {
    "schemas": {
...
      "BlockedWordSchema": {
        "properties": {
          "action": {
            "default": "BLOCK",
            "enum": [
              "BLOCK",
              "MASK"
            ],
            "title": "Action",
            "type": "string"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "keyword": {
            "maxLength": 200,
            "minLength": 1,
            "title": "Keyword",
            "type": "string"
          }
        },
        "required": [
          "keyword"
        ],
        "title": "BlockedWordSchema",
        "type": "object"
      },
...
      "CategoryRefSchema": {
        "properties": {
          "action": {
            "default": "BLOCK",
            "enum": [
              "BLOCK",
              "MASK"
            ],
            "title": "Action",
            "type": "string"
          },
          "category": {
            "maxLength": 100,
            "minLength": 1,
            "title": "Category",
            "type": "string"
          },
          "category_file": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Category File"
          }
        },
        "required": [
          "category"
        ],
        "title": "CategoryRefSchema",
        "type": "object"
      },
...
      "ContentPolicyCreateSchema": {
        "properties": {
          "active": {
            "default": false,
            "title": "Active",
            "type": "boolean"
          },
          "blocked_words": {
            "items": {
              "$ref": "#/components/schemas/BlockedWordSchema"
            },
            "title": "Blocked Words",
            "type": "array"
          },
          "categories": {
            "items": {
              "$ref": "#/components/schemas/CategoryRefSchema"
            },
            "title": "Categories",
            "type": "array"
          },
          "description": {
            "default": "",
            "maxLength": 500,
            "title": "Description",
            "type": "string"
          },
          "policy_id": {
            "maxLength": 64,
            "minLength": 1,
            "title": "Policy Id",
            "type": "string"
          }
        },
        "required": [
          "policy_id"
        ],
        "title": "ContentPolicyCreateSchema",
        "type": "object"
      },
      "ContentPolicyUpdateSchema": {
        "properties": {
          "active": {
            "default": false,
            "title": "Active",
            "type": "boolean"
          },
          "blocked_words": {
            "items": {
              "$ref": "#/components/schemas/BlockedWordSchema"
            },
            "title": "Blocked Words",
            "type": "array"
          },
          "categories": {
            "items": {
              "$ref": "#/components/schemas/CategoryRefSchema"
            },
            "title": "Categories",
            "type": "array"
          },
          "description": {
            "default": "",
            "maxLength": 500,
            "title": "Description",
            "type": "string"
          }
        },
        "title": "ContentPolicyUpdateSchema",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "PolicyCreateSchema": {
        "properties": {
          "ai_act_mode": {
            "title": "Ai Act Mode",
            "type": "boolean"
          },
          "entity_configs": {
            "additionalProperties": {
              "type": "string"
            },
            "title": "Entity Configs",
            "type": "object"
          },
          "gdpr_mode": {
            "title": "Gdpr Mode",
            "type": "boolean"
          },
          "headroom_mode": {
            "title": "Headroom Mode",
            "type": "boolean"
          },
          "is_active": {
            "title": "Is Active",
            "type": "boolean"
          },
          "name": {
            "title": "Name",
            "type": "string"
          }
        },
        "required": [
          "name",
          "is_active",
          "entity_configs",
          "gdpr_mode",
          "ai_act_mode",
          "headroom_mode"
        ],
        "title": "PolicyCreateSchema",
        "type": "object"
      },
      "PolicyUpdateSchema": {
        "properties": {
          "ai_act_mode": {
            "title": "Ai Act Mode",
            "type": "boolean"
          },
          "entity_configs": {
            "additionalProperties": {
              "type": "string"
            },
            "title": "Entity Configs",
            "type": "object"
          },
          "gdpr_mode": {
            "title": "Gdpr Mode",
            "type": "boolean"
          },
          "headroom_mode": {
            "title": "Headroom Mode",
            "type": "boolean"
          },
          "is_active": {
            "title": "Is Active",
            "type": "boolean"
          },
          "name": {
            "title": "Name",
            "type": "string"
          }
        },
        "required": [
          "name",
          "is_active",
          "entity_configs",
          "gdpr_mode",
          "ai_act_mode",
          "headroom_mode"
        ],
        "title": "PolicyUpdateSchema",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    }
  },
  "info": {
    "description": "Secure, white-labeled AI gateway with PII/PHI masking, budgets, and compliance policies.",
    "title": "AI Gateway API",
    "version": "1.0.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/content-policies": {
      "get": {
        "description": "Las políticas de contenido que existen hoy, con su estado real (activa o no).",
        "operationId": "list_content_policies_api_v1_content_policies_get",
        "parameters": [
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Content Policies",
        "tags": [
          "Content Policies"
        ]
      },
      "post": {
        "description": "Crea una política nueva. Aplica al siguiente pedido si `active=true`, sin reiniciar\nnada. (Las categorías que se cargan por fichero en el motor sí requieren un reinicio;\néstas no.)",
        "operationId": "create_content_policy_api_v1_content_policies_post",
        "parameters": [
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ContentPolicyCreateSchema"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Content Policy",
        "tags": [
          "Content Policies"
        ]
      }
    },
    "/api/v1/content-policies/{policy_id}": {
      "delete": {
        "description": "Borra una política. El filtro deja de aplicarse desde el siguiente pedido, sin\nreiniciar nada. La acción no se puede deshacer: para desactivarla temporalmente\nconservando sus reglas, usá el `PUT` con `active=false`.",
        "operationId": "delete_content_policy_api_v1_content_policies__policy_id__delete",
        "parameters": [
          {
            "in": "path",
            "name": "policy_id",
            "required": true,
            "schema": {
              "title": "Policy Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Delete Content Policy",
        "tags": [
          "Content Policies"
        ]
      },
      "put": {
        "description": "Actualiza (agregar/quitar palabras, activar/desactivar) una política existente.",
        "operationId": "update_content_policy_api_v1_content_policies__policy_id__put",
        "parameters": [
          {
            "in": "path",
            "name": "policy_id",
            "required": true,
            "schema": {
              "title": "Policy Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ContentPolicyUpdateSchema"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Update Content Policy",
        "tags": [
          "Content Policies"
        ]
      }
    },
...
    "/api/v1/security/policies": {
      "get": {
        "operationId": "list_policies_api_v1_security_policies_get",
        "parameters": [
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Policies",
        "tags": [
          "Security & Optimization"
        ]
      },
      "post": {
        "operationId": "create_policy_api_v1_security_policies_post",
        "parameters": [
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PolicyCreateSchema"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Policy",
        "tags": [
          "Security & Optimization"
        ]
      }
    },
    "/api/v1/security/policies/{policy_id}": {
      "delete": {
        "operationId": "delete_policy_api_v1_security_policies__policy_id__delete",
        "parameters": [
          {
            "in": "path",
            "name": "policy_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Policy Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Delete Policy",
        "tags": [
          "Security & Optimization"
        ]
      },
      "put": {
        "operationId": "update_policy_by_id_api_v1_security_policies__policy_id__put",
        "parameters": [
          {
            "in": "path",
            "name": "policy_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Policy Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PolicyUpdateSchema"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Update Policy By Id",
        "tags": [
          "Security & Optimization"
        ]
      }
    },
...
  }
}